changed type of int

+++ 0x1A-hash_tables/2-key_index.c
@@ -9,7 +9,7 @@
 
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int index, hash;
+	unsigned long int index, hash;
 
 	hash = hash_djb2((unsigned char *)key);
 	index = hash % size;
